introduce the polynomial authenticators for bitrot protection
This change adds a class of of new hash functions (polynomial authenticators)
to the existing hash-functions for bitrot protection.

Polynomial authenticators require a unique key. Therefore this change updates
the xl format to support either general purpose hash functions (without a key)
or poly. authenticators (with key).

This change affects only xl, not fs or gateway.