introduce the pkg/bitrot in the minio server
This change replaces the old bitrot code with the pkg/bitrot package.
It is the first step of moving from a hash function based bitrot detection
approach to an encryption & authentication based one.

Instead of using a hash.Hash interface for bitrot checksums it is now also
possible to use other detection techniques (like poly. authenticators).