encryption: replace bitrot verification skip- mechanism
This change replaces the bitrot verification skip-mechanism.
Before this change it was necessary to set the checksum of the bitrot-info
to nil/empty-string to indicate that the bitrot-verification should be skipped.
The problem with such kind of logic is that it is not obvious why/when certain code is
necessary. Especially when the relevant code-parts/functions are in separate files.
This is basically an implicit cross-function data/state dependency.

This change makes this explicit and encapsulates the state in the BitrotInfo struct.
Updates #4359