encryption: add general encryption support
This change adds general encryption support for creating/writing/healing
erasure-coded objects.
Therefore this change adds an additional secret key parameter which is passed
to createFile, readFile and healFile. This change also implements a specific version
of ChaCha20Poly1305.
This change only adds encryption support for the XL layer - it is currently not availabe from
the ObjectLayer/API.